feat(cli): remove need of rollups subcommand
