+++ .changeset/thick-taxis-pump.md
@@ -1,0 +1,5 @@
+---
+"@cartesi/cli": patch
+---
+
+remove need of rollups subcommand
