+++ apps/cli/src/commands/run.ts
@@ -4,11 +4,11 @@ import chalk from "chalk";
 export const createRunCommand = () => {
     return new Command("run")
         .description("Run a local cartesi node for the application.")
-        .summary("DEPRECATED: use 'rollups start' instead")
+        .summary("DEPRECATED: use 'start' instead")
         .action(async () => {
             console.warn(
                 chalk.yellow(
-                    "run command is deprecated, use 'rollups start' and 'rollups deploy' instead",
+                    "run command is deprecated, use 'start' and 'deploy' instead",
                 ),
             );
         });
