+++ apps/cli/src/compose/docker-compose-anvil.yaml
