+++ apps/cli/src/compose/docker-compose-bundler.yaml
