+++ apps/cli/src/compose/docker-compose-database.yaml
