+++ apps/cli/src/compose/docker-compose-explorer.yaml
