+++ apps/cli/src/compose/docker-compose-graphql.yaml
