+++ apps/cli/src/compose/docker-compose-node-cpus.yaml
