+++ apps/cli/src/compose/docker-compose-node-memory.yaml
