+++ apps/cli/src/compose/docker-compose-node.yaml
@@ -19,7 +19,7 @@ services:
             retries: 5
         command: ["cartesi-rollups-node"]
         env_file:
-            - ${CARTESI_BIN_PATH}/compose/rollups/default.env
+            - ${CARTESI_BIN_PATH}/compose/default.env
 
     proxy:
         volumes:
