+++ apps/cli/src/compose/docker-compose-paymaster.yaml
