+++ apps/cli/src/compose/docker-compose-proxy.yaml
