+++ apps/cli/src/compose/proxy/bundler.yaml
