+++ apps/cli/src/compose/proxy/graphql.yaml
